+++ ChangeLog.md
@@ -4,6 +4,12 @@ Project Coordinators: Florian Dubath @fdubath
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.4.3]
+### Changed
+- Add Gaussian and LogNormal Scaling Prior
+- Update the related section in the User Manual
+- Add Unit Tests for the scaling priors
+
 ## [1.4.2] 2024-11-04
 ### Changed
 - Update the Doc and move the QuickStart data in another project